requirements-core.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic>=2.7.0,<3.0.0
numpy>=1.26.0,<2.0.0
pandas>=2.1.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
pyyaml>=6.0.0,<7.0.0
httpx>=0.27.0,<1.0.0
pytest>=8.2.0,<9.0.0
pytest-asyncio>=0.23.8,<0.24.0
python-dateutil>=2.8.2,<3.0.0